Symbolicate Android framework frames
Summary:
This diff introduces storing a mapping in the Profilo trace from method ID to class_name + method_name for framework methods, because we can't symbolicate those by precomputing them at build time.

The idea is that, for every frame in a stack, we'll check - somewhat naively - if it's a framework method. If it is, we'll store a pointer to the class and method names. Then, the logger will write the mapping in the form of a `JAVA_FRAME_NAME` entry - with the method id as the `extra` field - followed by a `STRING_VALUE` entry with the concatenation of the class and method name as the value. Something like this:

```
-15|JAVA_FRAME_NAME|40299|2|0|0|0
307|STRING_VALUE|306|Ljava/lang/Runtime;nativeLoad
2|JAVA_FRAME_NAME|548|0|0|0|4110283702272
309|STRING_VALUE|308|Ljava/lang/reflect/Method;invoke
```

There's no need to have multiple equal entries, so we only write an entry for each method id once, which is enough to build the in-memory dictionary when parsing the trace.

Reviewed By: BurntBrunch

Differential Revision: D13536539

fbshipit-source-id: f1a34bf96ea3204dc1fd00f98ae96846df04d69d