Avoid building extra file name for every provider
Summary: Building extra file names for every provider have a cumulative cost in order of ~10 ms on a high-end device which can be avoided the file name will be generated on demand. As trace init is a critical time for tracing we can save cpu cycles by abstracting out the file generating into a provider.

Reviewed By: BurntBrunch

Differential Revision: D14135811

fbshipit-source-id: 11e17bd60443d2db688c93adda96dc2cf62cd72f