Check atrace symbol before hooking
Summary:
This diff has a goal of reducing time of a `SystraceProvider` init.

Not all dynamic libraries actually log to `ATRACE` so we can save some resources in `SystraceProvider` by checking a "magic" symbol and blacklisting the libraries which don't contain it. Thus we will be saving time on traversing plt section and also saving tracing runtime cpu as we won't be getting writes we surely know are not of our interest.

For "magic" symbol I chose `atrace_setup` which looks reliable indicator if a library references `ATRACE`. `atrace_marker_fd` could be another option but I found one instance where a library which actually logged into atrace didn't reference this symbol so I am going with the former one.

As part of this diff I refactored hooking code to factored `allowHookingCb` callback logic out of `profilo/util/hooks` and gave the clients to make appropriate implementation, so was I able to add additional filtering in `Atrace.cpp` without affecting `LibcIO` for example.

Reviewed By: ricardorey10

Differential Revision: D14211682

fbshipit-source-id: 8744cc4112ec3ba15002e32c90f1e66ecf8c594f