Store r7 in ip instead of on stack for ARMEABI syscalls
Summary:
Manually pushing r7 to the stack screws up unwind frames, because the unwind frame generated by the compiler becomes inaccurate and offset from the actual data. Mucking around with `.cfi_adjust_cfa_offset` and `.save` directives was unfruitful and attempts to use `LSS_REG` to control `r7` as a regular C variable - allowing the compiler to save/restore `r7` itself - actually ran into compiler bugs where it didn't save/restore anything and ended up clobbering frame pointers. The AAPCS allows free use of `ip` inside subroutines without requiring it to be saved and restored, so we can just save `r7` to `ip` and be job done.

This is effectively a re-land of an earlier version of D14306505.

Reviewed By: simpleton

Differential Revision: D14544961

fbshipit-source-id: c330d88d0a3d7a947a0cf1e724eb17b8f7f531e2