Trace stack collection errors
Summary:
This diff records the following stack collections errors:

- Empty stack
- Stack overflow
- No stack for "this" thread
- Signal interrupt: SIGSEGV or SIGBUS happened while unwinding)
- Nested unwind: SIGPROF was delivered while in the middle of unwinding

To achieve this, we simply reuse the `StackSlot` we were supposed to fill with the stack frames, and instead we use its state to record the error. On flush, we record the error type (one of the above) as the entry `type`, and the tracer type (Dalvik, ART_UNWINDC_X_Y_Z, etc...) as the `extra` field.

Initially, this approach resulted in a bunch of "missed_slot" errors due to the fact that, before this diff, on errors we would free the slot - allowing other threads to use it for their frames - , but now we were holding those slots until flush time, leading to data quality issues. To fix this, two approaches were taken:

- Ignore certain errors, like "tracer disabled" (which happens very often for JS tracer)
- Signal the reader thread more often (changed FLUSH_STACKS_COUNT from signaling every 8 stacks to every 4).

Reviewed By: BurntBrunch

Differential Revision: D14790890

fbshipit-source-id: 97b65c3c1d55b11f7e53314c81b603445fee2581