Make BytesEntries much cheaper
Summary:
Starting with Android 8, the concurrent copying collector triggers a different path in `GetStringCritical` - we now call `IncrementDisableThreadFlip` (http://androidxref.com/8.0.0_r4/xref/art/runtime/jni_internal.cc#1758).

This is a problem because any contention in that operation ends up measuring the duration of the stall by using `NanoTime()`, i.e. a syscall. This showed up in benchmarks and using `GetStringChars` (which almost always copies) is a solid 30% faster on average.

Reviewed By: tmelz

Differential Revision: D15270266

fbshipit-source-id: 28611c15a65f8178153791967ba3d11d4772e136