Test signal unregister edge cases
Summary:
This diff introduces tests for the main edge cases in signal unregistering.
It achieves this by virtue of `TestSequencer`, a poor man's `TurnSequencer`,
specifically meant to be used with an enum describing the states.

`TestSequencer` waits for a specific state and can then proceed to the next
one. This allows us to serialize execution (or rather, insert happens-before
relationships) to make hard situations explicit.

Reviewed By: ricardorey10

Differential Revision: D15425255

fbshipit-source-id: 109fece1cb20ce5eab85396a6130f0f903defc55