Scope signal handlers to an active profiling session
Summary:
The fault handler for SamplingProfiler is now pretty heavy - it has
to linearly scan every on every fault.

To mitigate this overhead, scope the handlers to the active profiling session.
This introduces some edge cases to worry about (pending signals and concurrent
handler execution for both fault and profiling) but it turns out, these are
very manageable.

Details on the edge cases and their mitigations are inline in the code.

Reviewed By: aandreyeu

Differential Revision: D15425254

fbshipit-source-id: 35125ccfc19021070b9b3339503bb52e40412c18