Align perf events to CLOCK_MONOTONIC
Summary:
Perf events come stamped from their own perf clock. Unfortunately, userspace
doesn't really have arbitrary access to this clock, we can only read it via
perf events.

Therefore, we devise a scheme where we read minor faults from a target thread,
intentionally trigger a minor fault on that thread, as well as capture the
monotonic time before and after it.

This allows us to average the monotonic time and compute a single offset from
the perf clock.

Is this perfect? Not really, we're assuming the perf clock is stable across
cores. However, we're gonna trust the kernel to equalize the differences fairly
reliably and not bother making this more clever without an actual need.

Reviewed By: aandreyeu

Differential Revision: D15990346

fbshipit-source-id: 175620737a648a46129b36350a1573dd94ac7e16