Constexprify the sample field offsets
Summary:
The perf_event_open API allows you to select different fields to be present in
samples. However, the fields are always ordered in the same way.
Therefore, to know the actual offset of a field, you need to know what was
requested first and skip the appropriate number of bytes.

The current logic calculates the offset from scratch for every field.
This is repetitive work and while the branch predictor will be on our side,
it's entirely unnecessary.

This diff does 3 things:

1. Fix the sample_type and read_format as constants that both Event and RecordSample can refer to.
2. Use constexpr to compute the offsets at compile time. This allows us to keep the hard-to-reason-about logic while losing nothing in terms of perf.
3. Switch to -O3 and observe all the magic inlining.

Reviewed By: ricardorey10

Differential Revision: D16431704

fbshipit-source-id: d7b0d8d44cdf3961e3a1ea8876c5d1fd4a32eed3