Optionally extend the trace past the "normal" trace end
Summary:
This diff implements the oft-requested feature of "trace past the end of the
trace." This is useful for instrumentation validation.

We achieve this by pushing out the TRACE_END event without touching the
TRACE_PRE_END (actual time that the end was requested).

This gives us a configurable time window that won't impact aggregation.

Reviewed By: aandreyeu

Differential Revision: D17237486

fbshipit-source-id: 5f07fa69d0a5123cc27bf70f5179ceff17d25ebd