Finer-grained locking
Summary:
Currently, we hold a lock for the whole duration of `getFileName()`, which is probably heavily-ish contended, as we call it from every single `replacement.*()` function. We shouldn't do this, so this diff introduces finer-grained locking.

The key observation is that in the "fast path" for `getFileName()` (namely, when the `fd` is either blacklisted or is already in our file cache) there are no writes to our global structures, so a read lock suffices. In the "slow path", that is, when we have to call `readlink()` to figure out the name of the symlink, we first release the read lock, do the lookup, and then we acquire it as a write lock only when we're going to touch our shared data.

This diff takes the read/write locks from [locks.h](https://fburl.com/diffusion/3f5r5eei) and extends them with the ability to `lock()` and `unlock()` them at will.

Reviewed By: tophyr, BurntBrunch

Differential Revision: D17375067

fbshipit-source-id: 898d92bebb5aa02c411df8f6fe1db143f7b7d729