Keep symmetry of TRACE_PRE_END trace
Summary:
There's a slight asymmetry with our trace lifecycle annotations.

Basically, when a trace starts, the `TRACE_START` entry is logged by the calling thread, not by one of Profilo's own threads. The same happens with aborts. However, for trace stops, the `TRACE_PRE_END` entry is logged asynchronously by the `TraceControlHandler` thread, which is wrong.

This diff takes care of that asymmetry.

Reviewed By: BurntBrunch

Differential Revision: D18099038

fbshipit-source-id: 8ff5d842afab521403a1641ce38750bb5301e378