Get rid of unnecessary constructor
Summary: Going through the code I noticed this 3-parameter constructor, which all it does is eventually call into the 2-parameter constructor, totally ignoring the first `nullptr`. I also noticed that in every caller we use `nullptr` as the first argument, which further made me believe that this constructor is useless.

Reviewed By: tophyr, aandreyeu

Differential Revision: D18739637

fbshipit-source-id: 90c61a2bf6409448095eebcd9a6a01ca2249ac39