Parse /proc/meminfo to get the size of the page cache
Summary:
This diff introduces parsing for /proc/meminfo.

`meminfo` is interesting because of two reasons:
1) vmstat is locked down on modern Android
2) it breaks down the active and inactive list sizes (parts of the page cache)

Therefore, we port counters that we used to take from vmstat to meminfo and add multiple new ones.

This diff also extracts a base class for files formatted in a similar way (i.e. ordered key-value pairs with left-padded values).

Reviewed By: yukonfb

Differential Revision: D17971104

fbshipit-source-id: 2357b527e9ac6fe3b971d56207953bf9e9824db0