Libraries whitelist for Atrace
Summary:
I've run some checks what are the actual libraries from which we get ATRACE blocks and came up with the list of 7 system libraries which log most of the time.
The theory is that the list of libraries is pretty much the same across multiple OS versions and instead of iterating over all the shared libraries we can just hook directly a small subset of 7 libraries which actually do the logging.
That can improve init timing for `SystraceProvider` on the OSes below 8.0 which are not covered by fast hook.
My local debugging shows we can expect 3 times improvement on average, but I am going to run an experiment to understand it better.

Reviewed By: ricardorey10

Differential Revision: D19955557

fbshipit-source-id: 8d6e9b9dd8b6d83afaf35f41b7c335ef9f787fc0