Hopefully fix SIGSEGV in distract_lookup_size while calling dlopen(null,)
Summary:
Some dalvik devices seem to not be able to correctly handle dlopen with null (which is a valid way to ask for the current lib) for some reason.

We already have another way to query for the shared libs which we do on Android L+. This diff just tries to do it on all Android versions if it can and it seems that
there is no reason this approach should not work on all versions. It seems the original code was written with the assumption not all versions of android have 'dl_iterate_phdr'
however from looking at the source code on dalvik, it appears it should, so lets try.

Differential Revision: D20079165

fbshipit-source-id: d151f0ab5c267a7a6d18c7efe643ea3025ebe2ec