feat(website): add DocSearch as recommended by docusaurus (#60)
Summary:
👋 team,

I am working on DocSearch. [We have an integration for Docusaurus](https://docusaurus.io/docs/en/search#enabling-the-search-bar). We thought it is a shame you can not also used this search that you can [find on reactjs for example](https://reactjs.org/).

This PR will add DocSearch to the documentation website. It will allow an user to have a learn-as-you-type experience by displaying results thanks to a dropdown in a live way.

Let me know if you need anything.
Pull Request resolved: https://github.com/facebookincubator/profilo/pull/60

Reviewed By: yukonfb

Differential Revision: D20504121

Pulled By: Fortisque

fbshipit-source-id: a934f91d7e5156fe162f7c6d93a775c57c31770c