Handle index_ for shorty
Summary:
I noticed that the art compatibility check was ALWAYS failing from the previous diff. Even though I had worked on D21090405 (to avoid using the dark greylist method dexMethodIndex in java stack trace) I noticed it still crashed in getting the c++ stack deep in unwindc.

By setting a breakpoint on the compatibility check in native code I was able to discover this code path which crashed. This small change allows us to symbolically unwind when the stack contains a native method.

I am pretty sure this was supposed to go on OS 8 as well based on the AOSP reference, so I've added it there as well.

Reviewed By: BurntBrunch

Differential Revision: D21156157

fbshipit-source-id: 00ea298ed2535d592f8417b903f837388f1dab28