Make atrace not crash on x86_64
Summary:
If a trace has the atrace provider enabled, but the platform doesn't match this supported platform regex it will cause the app to crash.

For example trying to collect a manual trace on x86_64 will cause the app to crash because manual traces have all providers enabled. The app would crash with `java.lang.UnsatisfiedLinkError: couldn't find DSO to load: libprofilo_atrace.so`

This diff removes the supported platform regex so that it's possible to collect manual traces on x86_64

Reviewed By: BurntBrunch

Differential Revision: D21388727

fbshipit-source-id: c18ce0db97d247d07688708f9f3106cc9f2c21d1