Fix pedantic initializer warnings
Summary:
Nested initialization is apparently a C99-only feature.
C++ only supports one level of initializers.

The sigaction case is interesting because `sa_handler` and `sa_sigaction` are
not defined as being in a union but in practice they are. To avoid using the
name of this union member, we use default initialization instead.

Reviewed By: aandreyeu

Differential Revision: D23478107

fbshipit-source-id: 159edf02cf3c5a5c5ab594a2ce3b508d6a30c761