Introduce SignalHandler as a lightweight alternative to sigmux
Summary:
There are a couple of problems with our usage of sigmux.

1. If we do in fact are the first library to initialize `sigmux`, we'd end up installing a `sigaction` wrapper which puts `sigmux` entirely in front of `sigchain` (and art's fault manager). This is in line with our goals and what we use `sigmux` for.
a. Caveat: on Android 9+ this doesn't actually happen because of the distinction between `sigaction` and `sigaction64`. Ignore that for now.
b. However, if we put `sigmux` in front of `sigchain` and someone *else* (e.g. fb-breakpad) registers with `sigmux`, we may break *their* expectations. For example, breakpad would run *before* art's fault handler, thus generating minidumps for recoverable faults (e.g. NPEs).
2. If we are *not* the first library to initialize sigmux, then we are likely running *after* art's fault handling (as we can't assume that anyone has put sigmux in front of sigchain). This should be fine but isn't - art's fault handler can get confused and dereference garbage, causing a double fault and a consecutive crash.

In this commit we introduce a profilo-specific solution for signal handling for the sampling profiler. It completely ignores sigmux and operates at the basic `sigaction` layer (by using libc's `sigaction` or `sigaction64` directly).

We implement a much simpler and restricted version of sigmux:
1) We can look up a specific `SignalHandler` from an execution of its actual signal handler function.
2) We can wait for all active signals to finish processing (by using sigmux's own `phaser_t`)
3) We can call the previous signal handler for this signal.

We cannot remove a `SignalHandler` once installed as `SignalHandler` does not itself gain control of the signal when executing, it always passes the `HandlerPtr` to the `sigaction` call. The installed handlers are responsible for communicating with `SignalHandler` via the `EnterHandler` call and the corresponding methods on `HandlerScope`.

Reviewed By: aandreyeu

Differential Revision: D23248677

fbshipit-source-id: 9fa0fd483576ef3710fd1b21b4c464c093968788