Avoid temporaries in toCall
Summary:
The old versions of these functions should look sketchy.  A local_ref
gets constructed for the pass-by-value argument, then we return the raw
jobject right before the local_ref is destroyed.  (alias_ref is fine;
I just converted it for consistency.)

The only reason we get away with this is that this method is always used
like `CallMethod(Convert::toCall(something))`.  The C++ standard specifies
"It is implementation-defined whether the lifetime of a parameter ends
when the function in which it is defined returns or at the end
of the enclosing full-expression."  And the Itanium C++ ABI specifies
"If the type has a non-trivial destructor, the caller calls that destructor
after control returns to it (including when the caller throws an exception),
at the end of enclosing full-expression."  So we're safe.  The temporary
local_ref survives until after CallMethod completes.

Except on Windows.  MSVC doesn't follow the Itanium ABI.  Its ABI requires
the callee to invoke the destructor for temporary values constructed for
arguments.  More discussion at
https://quuxplusone.github.io/blog/2018/11/12/parameter-lifetime-and-trivial-abi/

The fix is just to take the arguments by const-reference, so no temporary is
constructed.  This should be marginally more efficient as well.

Reviewed By: cjhopman

Differential Revision: D23862672

fbshipit-source-id: b26cfff76d864a3a58934f06d1da8f2b66668a11