Skeleton for deallocating buffers
Summary:
This diff restores the ability to deallocate buffers and hooks it up so all paths call deallocateBuffer.

Do note that TraceOrchestrator now maintains two different sets of trace objects - TraceContext and Trace objects. In particular, when receiving a callback from the background thread, we need to find the corresponding TraceContext, so we can deallocate the buffer correctly.

This is annoying and needs unification but that's a battle for another day.

Reviewed By: aandreyeu

Differential Revision: D23847732

fbshipit-source-id: 7ba76f0e186fda52fc42114643a9e77365d72982