Notify TraceControl of abort on malformed condition
Summary:
BurntBrunch noticed a few days ago that when a config is malformed we abort the trace on the handler, but `TraceControl` is never notified, so the trace won't end.

This diff changes the handler's code to call into `TraceControl` when we abort a trace, so that the correct steps are taken.

Reviewed By: BurntBrunch

Differential Revision: D25314657

fbshipit-source-id: acf1651ead4660e9dc4012a0ede966a1d3657399