Disable -Wreorder-init-list for sigaction usages
Summary:
Profilo sees conflicting definitions of `sigaction` that lists the
members in different orders and thus no one field designator order for
designated initilizers can satisfy all usages. Thus we just have to disable this
new warning here. Though until llvm 11 is rolled out to 100% of builds this will
fail with a unknown-warning error. So we must add that flag as well.

Reviewed By: smeenai

Differential Revision: D25620960

fbshipit-source-id: c75a4fea6d0e6bc12e9ae0393cd2a1513d22de61