Logger does not keep the entry ID count itself
Summary:
Instead of keeping the entry ID count as a private member of Logger, introduce a global instance of `EntryIDCounter` which is shared by the loggers in `mmapbuf::Buffer` instances.

`TraceWriters` which need to overwrite the starting entry ID can instantiate their own `Logger` with a new counter.

In effect, this ensures that buffers share entry IDs and therefore the same entry has the same ID across all concurrent traces.

Reviewed By: aandreyeu

Differential Revision: D25768234

fbshipit-source-id: db7ae6c13b402968c6cd01d29193094d2969e3ec