TraceContext and NativeTraceWriter take the exact trace folder as path
Summary:
As a remnant from the `MultiTraceLifecycleVisitor` days, `NativeTraceWriter`
takes the general trace folder (files/profilo) as a path and creates per-trace
folders inside of it.

Now that there's a 1-1 relationship between `NativeTraceWriter`s and traces, we
can just pass the concrete trace folder path to it. It still creates the folder
but not the path.

Most importantly, `TraceContext` now knows where the trace data is, which will come
in handy later on when we need to add files that do not have trace lifecycle
events.

Reviewed By: aandreyeu

Differential Revision: D26954450

fbshipit-source-id: 35f497a94007af15d736119c7f7d617159ec0539