Use TraceContext.folder instead of TraceIdToTrace HashMap
Summary:
`TraceOrchestrator` has two maps to help it navigate all the callbacks -
from trace ID to `TraceContext `and from trace ID to `Trace `objects, which hold
the exact trace log file.

Now that `TraceContext` has a `folder` field, we don't need the latter map and
most of the `Trace` class.

Reviewed By: aandreyeu

Differential Revision: D26954776

fbshipit-source-id: b750615bf2c473125879e01490d0b5fdb944e0e4