Aggregate mappings as system counters
Summary:
Memory shared with hardware devices shows up as named mappings and is not generally countable via predefined counters. There is Debug.getMemoryInfo but that parses smaps instead of just maps and doesn't count all the things we care about.

This diff introduces a simple mapping aggregator that starts off only counting the size of `dmabuf` and `kgsl` mappings.

Reviewed By: aandreyeu

Differential Revision: D28806833

fbshipit-source-id: 98f6316174fe03fd9dd86e6f55df00f7948b5ab8