Use /proc/self/mem writes as a faster way to patch GOT slots
Summary:
On older versions of Android (pre-24), the kernel follows the mainline behavior and allows writes to `/proc/self/mem`. What's interesting about this functionality is that these writes bypass PROT_* flags on the vma, so you can directly change an `r--` page.

In this diff, we implement the most naive version of this alternative patching logic - opening and closing the file for every write - and nevertheless observe significant speedups.

Reviewed By: aandreyeu

Differential Revision: D27482717

fbshipit-source-id: 42153cdfd248002e901749c6cd49dae449961fbe