Fix up more gen_aidl rules
Summary:
Similar to previous diff, need these in order to build them successfully with V2 since we need to send the `aidl_srcs` to RE.

We also remove some rules that are not needed because on V2 we don't produce an artifact for these rules and V2 doesn't like that. In V2 we just create an empty zip file.

Reviewed By: mykola-semko

Differential Revision: D32202636

fbshipit-source-id: 80cdde685841d257fa54cf9f4e46e56a1392f99f