SignalHandler can change registered handlers safely
Summary:
SignalHandler currently does not allow the registered handler for a signal to
change. I want to use SignalHandler in a different, non-overlapping with
SamplingProfiler, use case and to do that, I need to be able to safely change
the registered handler.

This diff achieves this in three steps:
1. Instead of registering the user-provided signal handler directly with libc,
we register UniversalHandler instead. This way, once we take over the signal
slot, we don't have to call libc again for that signal.

2. This diff also wraps the access to the global array of user-provided handlers in a
phaser. This way, we can wait until concurrent users of the old SignalHandler
instance are done.

3. Lastly, we propagate the "previous handler" that we got from sigaction(3)
from the old SignalHandler instance to the new one. This way, we never
chain our own SignalHandler instances together but instead they always refer to
what was there before the first SignalHandler took over.

Reviewed By: aandreyeu

Differential Revision: D32322550

fbshipit-source-id: 8ea6cc8b16e60fc1d6ffdedb2cd4a03d6b42c3cc