docs: add social banner in support of Ukraine (#108)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.
<img width="1080" alt="image" src="https://user-images.githubusercontent.com/12485205/156703199-25bb934a-fd36-4220-b687-805f68526f2a.png">

Pull Request resolved: https://github.com/facebookincubator/profilo/pull/108

Reviewed By: aandreyeu

Differential Revision: D34634750

Pulled By: dmitryvinn-fb

fbshipit-source-id: 9cc5e95152e2b48dc505f9733ccfb92c4b5d0792