Reversible Trace ID sanitization
Summary:
# Problem Description

Loom uses base64 encoding for the string representation of Trace IDs. The base64 alphabet uses `+` and `/` characters in addition to alphanumeric characters. Trace file names contain Trace IDs, and in order it to be valid filenames, Trace IDs are sanitized to replace unsupported characters.

The current approach is replacing both `+` and `/` with an underscore (`_`). This transformation makes it impossible to reconstruct the actual Trace ID from the filename, which could be useful in situations where we need to access the Trace ID without parsing the trace file contents (on either client or server).

# Change Description

With this change, the sanitization function becomes reversible, by having it replace `+` with `_p_` and `/` with `_s_`.

Reviewed By: aandreyeu

Differential Revision: D36764316

fbshipit-source-id: 744ac37fe6c190436a2445170fac6d1003b3b2f1