Back out "Temporarily disable stack traces for libc++ arm64"
Summary:
As with all temporary solutions, this turned out to be a lot less
temporary than I was hoping for :D I was never able to figure out the
original issue, but I tried again after switching unwinders from libgcc
to libunwind. Unfortunately that ran into new crashes so I had to revert
it again. I fixed the underlying issue in LLVM's libunwind in
https://reviews.llvm.org/D126343, so let's try this again and see what
happens.

Reviewed By: simpleton

Differential Revision: D37181295

fbshipit-source-id: bb4ca5c8e525e8deab3fa43b8be78120b477ffed